refractor app sbatch options message in gRPC
