+++ .changeset/khaki-mugs-try.md
@@ -1,0 +1,7 @@
+---
+"@scow/portal-server": patch
+"@scow/portal-web": patch
+"@scow/grpc-api": patch
+---
+
+重构 app 的 sbatch options，gRPC 中与 custom_attributes 一起发送
