+++ protos/portal/app.proto
@@ -59,7 +59,6 @@ message CreateAppSessionRequest {
   uint32 core_count = 7;
   uint32 max_time = 8;
   map<string, string> custom_attributes = 9;
-  repeated string user_sbatch_options = 10;
 }
 
 // NOT_FOUND: app is not found
